backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# DB
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Cache / 세션
redis==5.1.1

# 통계 엔진
scipy==1.14.1
numpy==2.1.2

# 파일 파싱
openpyxl==3.1.5
pandas==2.2.3

# 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 유효성 검증
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12

# HTTP 클라이언트 (ClamAV 연동)
httpx==0.27.2

# 보고서 생성
weasyprint==62.3

# 유틸
python-dotenv==1.0.1
structlog==24.4.0
